+++ draft-js-mention-plugin/src/index.js
@@ -9,12 +9,14 @@ import mentionSuggestionsStrategy from './mentionSuggestionsStrategy';
 import suggestionsFilter from './utils/defaultSuggestionsFilter';
 import defaultPositionSuggestions from './utils/positionSuggestions';
 import { defaultTheme } from './theme.js';
+import addMention from './modifiers/addMention';
 
 export {
   default as MentionSuggestions,
 } from './MentionSuggestions/MentionSuggestions';
 
 export { defaultTheme };
+export { addMention };
 
 export default (config = {}) => {
   const callbacks = {
@@ -85,6 +87,7 @@ export default (config = {}) => {
     mentionTrigger = '@',
     mentionRegExp = defaultRegExp,
     supportWhitespace = false,
+    entryComponent,
   } = config;
   const mentionSearchProps = {
     ariaProps,
@@ -95,6 +98,7 @@ export default (config = {}) => {
     positionSuggestions,
     mentionTrigger,
     mentionPrefix,
+    entryComponent,
   };
   const DecoratedMentionSuggestionsComponent = props => (
     <MentionSuggestionsComponent {...props} {...mentionSearchProps} />